+++ sulong/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 New features:
 
+* Support locating dynamic libraries relatively using (`rpath`).
 * Preliminary support for compiling to bitcode using the LLVM toolchain.
   See [docs/TOOLCHAIN.md](docs/TOOLCHAIN.md) for more details.